+++ OpenccNet/ConvertCommand.cs
@@ -120,7 +120,7 @@ internal static class ConvertCommand
             {
                 try
                 {
-                    CliUtils.ParseCustomDictSpec(value);
+                    CustomDictSpec.Parse(value);
                 }
                 catch (ArgumentException ex)
                 {
@@ -209,7 +209,7 @@ internal static class ConvertCommand
                 var dict = DictionaryLib.New();
 
                 var specs = customDicts
-                    .Select(CliUtils.ParseCustomDictSpec)
+                    .Select(CustomDictSpec.Parse)
                     .ToArray();
 
                 DictionaryLib.WithCustomDicts(dict, specs);