+++ OpenccNet/DictgenCommand.cs
@@ -57,7 +57,7 @@ internal static class DictgenCommand
             {
                 try
                 {
-                    CliUtils.ParseCustomDictSpec(value);
+                    CustomDictSpec.Parse(value);
                 }
                 catch (ArgumentException ex)
                 {
@@ -119,7 +119,7 @@ internal static class DictgenCommand
                     Console.WriteLine($"Applying {customDicts.Length} custom dictionary spec(s)...");
 
                     var specs = customDicts
-                        .Select(CliUtils.ParseCustomDictSpec)
+                        .Select(CustomDictSpec.Parse)
                         .ToArray();
 
                     DictionaryLib.WithCustomDicts(dict, specs);