+++ OpenccNet/OfficeCommand.cs
@@ -90,7 +90,7 @@ internal static class OfficeCommand
             {
                 try
                 {
-                    CliUtils.ParseCustomDictSpec(value);
+                    CustomDictSpec.Parse(value);
                 }
                 catch (ArgumentException ex)
                 {
@@ -154,7 +154,7 @@ internal static class OfficeCommand
                     var dict = DictionaryLib.New();
 
                     var specs = customDicts
-                        .Select(CliUtils.ParseCustomDictSpec)
+                        .Select(CustomDictSpec.Parse)
                         .ToArray();
 
                     DictionaryLib.WithCustomDicts(dict, specs);